Cope with requests with bad date formats
When we get a request with a date like "December 20, 2019",
we get an indexing error, because our csv now has one extra
comma than it should.
Now we try and cope with that by checking if it looks like
a date in a bad format, and if it does, we interpret the comma
as part of the value instead of a separator.